+++ .rspec
@@ -1,1 +1,2 @@
 --colour
+--require spec_helper
