+++ spec/controllers/home_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 RSpec.describe HomeController do
   describe 'GET #index' do
