+++ spec/controllers/profiles_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ProfilesController, :type => :controller do
   include HelperMethods::Controllers
