+++ spec/features/comment_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Comment", %q{
   In order to talk to other people
