+++ spec/features/events_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Events", %q{
   In order to open call for papers
