+++ spec/features/profiles_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature 'Profiles' do
   context 'showing user profile' do
