+++ spec/features/proposal_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Proposal", %q{
   In order to select proposals
