+++ spec/features/registration_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Registration", %q{
   In order to interact with the site
