+++ spec/features/vote_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Vote", %q{
   In order to express my opinion
