+++ spec/helpers/application_helper_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ApplicationHelper, :type => :helper do
   describe "#menu" do
