+++ spec/helpers/proposals_helper_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ProposalsHelper, :type => :helper do
   let(:user) { FactoryBot.create(:user) }
