+++ spec/helpers/user_helper_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 RSpec.describe UserHelper, type: :helper do
   context '#user_picture' do
